+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Remove `stream.py` and assume connections are not closed on return.
+
 ## [v0.9.4] - 01.10.2025
 
 ### Changed
@@ -23,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `url_link` in tools that retrieve explore data.
 - Remove `sonata_asset_id` from circuit analysis tool dependency.
+
 
 ## [v0.9.2] - 25.09.2025